No more header slideout for desktop > 800 px
Would have preferred to use a CSS media query to change an attribute to be
queried in the JS, but it's so much easier to make a query in JS at that point I guess.

NOTE: revert this change if you change your mind. Previous changes in this
branch were just cleanup.